Rename parser/scanner to operation-parser/scanner
